Created TextInput for article editor component
